pyproject.toml
[tool.poetry]
name = "ai_pocs"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
gradio = "^4.31.2"
openai = "^1.30.1"
python-dotenv = "^1.0.1"
pdf2image = "^1.17.0"
azure-ai-documentintelligence = "^1.0.0b3"
tiktoken = "^0.7.0"
langchain = "^0.2.15"
bs4 = "^0.0.2"
langchain-community = "^0.2.14"
pymupdf = "^1.24.9"
fitz = "^0.0.1.dev2"
langchain-openai = "^0.1.8"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-xdist = "^3.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
